ms-cliente/src/main/java/com/fase5/techchallenge/fiap/mscliente/usecase/endereco/ObterEnderecosPeloCliente.java
package com.fase5.techchallenge.fiap.mscliente.usecase.endereco;

import com.fase5.techchallenge.fiap.mscliente.entity.cliente.gateway.ClienteGateway;
import com.fase5.techchallenge.fiap.mscliente.entity.cliente.model.Cliente;
import com.fase5.techchallenge.fiap.mscliente.entity.endereco.gateway.EnderecoGateway;
import com.fase5.techchallenge.fiap.mscliente.entity.endereco.model.Endereco;
import com.fase5.techchallenge.fiap.mscliente.usecase.exception.BussinessErrorException;
import lombok.AllArgsConstructor;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;

@Service
@AllArgsConstructor
public class ObterEnderecosPeloCliente {

    private final ClienteGateway clienteGateway;
    private final EnderecoGateway enderecoGateway;

    public List<Endereco> execute(String idCliente) {
        Optional<Cliente> clienteOptional = clienteGateway.findById(idCliente);

        if (clienteOptional.isEmpty()) {
            throw new BussinessErrorException("Cliente não existe.");
        }
        return this.enderecoGateway.findByCliente(clienteOptional.get());
    }
}
